Moved the Out of the Box Experience to a separate class to: a) easy run in async mode, b) don't have to use static variables
